feat(console,cloud): full chain alignment — unified 4-type timeline with stats
Align Console order history with RedCoral's chain_entry model, supporting
ORDER, CREDIT_NOTE, ANULACION, and UPGRADE entry types end-to-end.

Cloud: expand list_chain_entries to 4-UNION ALL, add anulacion/upgrade
detail endpoints and query functions.

Console: rewrite OrdersScreen with type-aware rendering, icon system,
jump navigation between related entries, amount statistics bar
(sales - refunds = net, excluding voided orders), and human-readable
formatting for refund methods and AEAT status.

Backend: align sync types, invoice models, and archiving services for
full chain entry coverage.

+++ edge-server/src/archiving/invoice.rs
@@ -141,7 +141,7 @@ impl InvoiceService {
         for d in desglose {
             let base = decimal_to_f64(d.base_amount, "desglose.base_amount")?;
             let tax_amt = decimal_to_f64(d.tax_amount, "desglose.tax_amount")?;
-            inv_repo::insert_desglose(tx, invoice_id, i64::from(d.tax_rate), base, tax_amt).await?;
+            inv_repo::insert_desglose(tx, invoice_id, d.tax_rate, base, tax_amt).await?;
         }
 
         // Update system_state.last_huella within the same transaction
@@ -268,7 +268,7 @@ impl InvoiceService {
         for d in desglose {
             let base = decimal_to_f64(d.base_amount, "desglose.base_amount")?;
             let tax_amt = decimal_to_f64(d.tax_amount, "desglose.tax_amount")?;
-            inv_repo::insert_desglose(tx, invoice_id, i64::from(d.tax_rate), base, tax_amt).await?;
+            inv_repo::insert_desglose(tx, invoice_id, d.tax_rate, base, tax_amt).await?;
         }
 
         // Update system_state.last_huella within the same transaction
